Add CACERT variables for env subcommand
This PR introduces the <product>_CACERT variable in the `env` subcommand output.

Shikari assumes that the CA cert for the cluster is placed in the first servers instance directory in the name `<product>-agent-ca.pem`.

The responsiblity of placing the CA Cert in the instance directory is up to the respective scenarios.

The `copyToHost` functionality in Lima can be used for the same.